THE CREATE WEB API CALL IS NOW WORKING
The create flashcard API call is now working. I added a check to see if
the same flashcard already exists. It already exists if the English
and Russian words are the same as an existing flashcard.

I updated the flashcard model. It was using the User schema. The
model now is complete (for now) for a flashcard.

db.js now points to the russian-flashcards database in MongoDB. It was
pointing to a different database.

I added the DuplicateFlashcardError to custom_errors.js, and the
handleDuplicateFlashcardError() method as well that can be invoked
from the flashcard routes for handling a duplicate flashcard.

I added http-status-codes.js to the lib folder. This acts as an
enumeration for return codes so as to eliminate magic numbers.

I added the DuplicateFlashcardError to the errors to be processed
in error_handler.js It uses http-status-codes.js.

I added flashcard_routes.js to implement the routes for flashcards.

I added my flashcard routes to server.js.

I added the create.sh CURL script for creating a flashcard in the
curl-scripts/routes folder.

I added command-line.txt to have the command line invocations for
flashcards so that I cna copy and paste them to the terminal. It is a
pain to type them.