THE UPDATE WEB API CALL FOR FLASHCARD IS WORKING
I added the commandline invocation for update to command-line.txt and
added the CURL script.

I added the show route to flashcard_routes.js.

I was receiving exceptions about an ID not being found for update. I
debugged and realized that custom_error.js has a requiresOwnership
method for updates so that you cannot change the owner. In my schema,
I had flashcardOwner, but the method just mentioned wants the user
field to be "owner". Once I changed this it works.

I changed "flashcardOwner" to simply "owner" in my flashcard model.