THE DELETE WEB API CALL FOR FLASHCARD IS WORKING
I added the commandline invocation for delete to command-line.txt and
added the CURL script.

I added the delete route to flashcard_routes.js.

This is the last of the routes that I need to meet the minimum
requirements. I am sure that I'll find issues when I start on the
front end.